requirements.txt
clearml
numpy
tqdm
pillow
pandas
tensorboard
fightingcv-attention
timm
pytorch_wavelets
